Refactor: 이미지 처리 과정 순서 변경
발행 버튼 클릭 시 -> 이미지 추가 시 파이어베이스 주소로 변경
