+++ src/pages/editor/components/editor/hooks/useProcessContent.ts
@@ -1,4 +1,3 @@
-import { uploadImageToFirebase } from '@/hooks/useImageConvert';
 import { useRecoilValue } from 'recoil';
 import { editorState } from '../atoms/editor.atom';
 
@@ -10,25 +9,6 @@ const useProcessContent = () => {
 
     const parser = new DOMParser();
     const document = parser.parseFromString(content, 'text/html');
-
-    // TODO: 에러 핸들링 과정 추가
-    const images = document.querySelectorAll('img');
-    const imageUploads = Array.from(images).map(async (img, index) => {
-      console.log(`Processing image ${index}`, img.src);
-
-      try {
-        if (img.src.startsWith('https://')) return;
-        const blob = await fetch(img.src).then((res) => res.blob());
-        const firebaseUrl = await uploadImageToFirebase(blob);
-        console.log('Image uploaded. Firebase URL: ', firebaseUrl);
-        img.src = firebaseUrl;
-      } catch (error) {
-        console.error('Error Uploading image:', img.src, error);
-      }
-    });
-
-    await Promise.all(imageUploads);
-    console.log('All image uploads completed');
 
     const titleEl = document.querySelector('h1');
